+++ NEWS.rst
@@ -1,3 +1,11 @@
+RELEASE 3.0.1, Upcoming
+-----------------------
+
+bundle_adjust:
+  * Add option --apply-initial-transform-only to apply 
+    an initial trasform to camera while skipping
+    image matching and other steps. 
+
 RELEASE 3.0.0, July 27, 2021
 ----------------------------